requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==1.0.0
Django==3.2
django-cors-headers==3.6.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
